Fix class name

+++ program/skeleton/src/Entity/EV3/Devise/InputDevise/Sensor/LightSensor.java
@@ -1,15 +1,19 @@
 package Entity.EV3.Devise.InputDevise.Sensor;
 
-public final class LightSensor extends Sensor {
-  public LightSensor(char port) {
+public final class LightSensor extends Sensor
+{
+  public LightSensor(char port)
+  {
     super(port);
   }
 
-  public int getValue() {
+  public int getValue()
+  {
     return 0;
   }
   
-  public float getFloatValue() {
-    return (float)0.0;
+  public float getFloatValue()
+  {
+    return 0.0F;
   }
 }
